backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
redis==5.0.1
google-cloud-storage==2.13.0
google-cloud-vision==3.4.5
google-api-python-client==2.107.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
tenacity==8.2.3
bcrypt==4.0.1
supabase==2.3.4 